ocr-worker/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
pillow==10.4.0
numpy==1.26.4
paddleocr==2.7.0.3
paddlepaddle==2.6.2
